Apply xo formatting
Apply markdown formatting guidelines
