+++ README.md
@@ -1,10 +1,6 @@
-# Javascript DynamoDB Context Store 
+# Javascript DynamoDB Context Store
 
-Used by the [SmartApp SDK](https://github.com/SmartThingsCommunity/smartapp-sdk-nodejs) to store IDs and access tokens for an installed instance of a SmartApp
-and retrieves that information for use in asynchronous API calls. The use of a context store
-is only needed when SmartApps have to call the SmartThings API in response to external
-events. SmartApps that only response to lifecycle events from the SmartThings platform
-will automatically have the proper context without the app having to store it. 
+Used by the [SmartApp SDK](https://github.com/SmartThingsCommunity/smartapp-sdk-nodejs) to store IDs and access tokens for an installed instance of a SmartApp and retrieves that information for use in asynchronous API calls. The use of a context store is only needed when SmartApps have to call the SmartThings API in response to external events. SmartApps that only response to lifecycle events from the SmartThings platform will automatically have the proper context without the app having to store it.
 
 The context stored by this module consists of the following data elements:
 
@@ -17,8 +13,9 @@ The context stored by this module consists of the following data elements:
 _Note: Version 2.X.X is a breaking change to version 1.X.X as far as configuring the context store is
 concerned, but either one can be used with any version of the SmartThings SDK._
 
-## Installation:
-```
+## Installation
+
+```bash
 npm install @smartthings/dynamodb-context-store
 ```
 
@@ -26,11 +23,13 @@ npm install @smartthings/dynamodb-context-store
 
 Create a `DynamoDBContextStore` object and pass it to the SmartApp connector to store the context in a table 
 named `"smartapp"` in the `us-east-1` AWS region. If the table does not exist it will be created.
+
 ```javascript
 smartapp.contextStore(new DynamoDBContextStore())
 ```
 
 The more extensive set of options are shown in this example:
+
 ```javascript
 smartapp.contextStore(new DynamoDBContextStore(
     {
@@ -48,6 +47,7 @@ smartapp.contextStore(new DynamoDBContextStore(
 ```
 
 The **table** configuration options are:
+
 * **name** -- The name of the DynamoDB table storing the context
 * **hashKey** -- The name of the partition key of the table
 * **prefix** -- A string pre-pended to the installed app ID and used as the partition key for the entry
@@ -56,6 +56,7 @@ The **table** configuration options are:
 * **sortKey** -- Optional sort key definition (see below for more details)
 
 Other configuration options are:
+
 * **AWSRegion** -- The AWS region containing the table
 * **AWSConfigPath** -- The location of the AWS configuration JSON file
 * **AWSConfigJSON** -- The AWS credentials and region
@@ -67,6 +68,7 @@ Note that only one of the AWS options should be specified or behavior will be in
 
 By default, the AWS credentials are picked up from the environment. If you prefer you can read the credentials
 from a file with this configuration:
+
 ```javascript
 smartapp.contextStore(new DynamoDBContextStore(
     {
@@ -75,8 +77,8 @@ smartapp.contextStore(new DynamoDBContextStore(
 ))
 ```
 
-
 You can also explicitly set the credentials in this way:
+
 ```javascript
 smartapp.contextStore(new DynamoDBContextStore(
     {
@@ -87,11 +89,13 @@ smartapp.contextStore(new DynamoDBContextStore(
         }
     }
 ))
-````
+```
 
 ### Sort Key Configuration
+
 In order to support single table schemas, the context store can be configured to use a table with a sort key.
 The simplest way to do that is by specifying the sort key name:
+
 ```javascript
 smartapp.contextStore(new DynamoDBContextStore(
     {
@@ -102,10 +106,11 @@ smartapp.contextStore(new DynamoDBContextStore(
         }
     }
 ))
-
 ```
+
 More control over the sort key can be exercised using this form, which is configured with the default values
 used when just the sort key name is specified:
+
 ```javascript
 smartapp.contextStore(new DynamoDBContextStore(
     {
@@ -121,5 +126,4 @@ smartapp.contextStore(new DynamoDBContextStore(
         }
     }
 ))
-
 ```
